refactor(webapp-game): move off contribute owned projects move to actionMoves
