+++ packages/game/src/index.ts
@@ -9,7 +9,7 @@ import forceCards from './data/card/forces.json';
 import eventCards from './data/card/events.json';
 import { isInRange } from './utils';
 import { ActiveProject, ActiveProjects } from './activeProjects';
-import { createProject, recruit } from './moves/actionMoves';
+import { contributeOwnedProjects, createProject, recruit } from './moves/actionMoves';
 
 type PartialBy<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
 type WithGameState<G extends any, F extends (...args: any) => void> = (G: State<G>['G'], ctx: State<G>['ctx'], ...args: Parameters<F>) => any;
@@ -125,50 +125,10 @@ export const OpenStarTerVillage: Game<type.State.Root> = {
             client: false,
             move: recruit,
           },
-          contributeOwnedProjects: ((G, ctx, contributions) => {
-            if (!G.table.activeMoves.contributeOwnedProjects) {
-              return INVALID_MOVE;
-            }
-
-            const currentPlayer = ctx.playerID!;
-            const currentPlayerToken = G.players[currentPlayer].token;
-            const contributeActionCosts = 1;
-            if (currentPlayerToken.actions < contributeActionCosts) {
-              return INVALID_MOVE;
-            }
-            const activeProjects = G.table.activeProjects
-            const isInvalid = contributions.map(({ activeProjectIndex, jobName }) => {
-              if (!isInRange(activeProjectIndex, activeProjects.length)) {
-                return true;
-              }
-              const activeProject = ActiveProjects.GetById(activeProjects, activeProjectIndex);
-              if (activeProject.owner !== currentPlayer) {
-                return true;
-              }
-
-              if (!ActiveProject.HasWorker(activeProject, jobName, currentPlayer)) {
-                return true;
-              }
-            }).some(x => x);
-            if (isInvalid) {
-              return INVALID_MOVE;
-            }
-            const totalContributions = contributions.map(({ value }) => value).reduce((a, b) => a + b, 0);
-            const maxOwnedContributions = 4;
-            if (!(totalContributions <= maxOwnedContributions)) {
-              return INVALID_MOVE;
-            }
-
-            // deduct action tokens
-            currentPlayerToken.actions -= contributeActionCosts;
-            contributions.forEach(({ activeProjectIndex, jobName, value }) => {
-              // update contributions to given contribution points
-              const activeProject = ActiveProjects.GetById(G.table.activeProjects, activeProjectIndex);
-              ActiveProject.PushWorker(activeProject, jobName, currentPlayer, value);
-            });
-
-            G.table.activeMoves.contributeOwnedProjects = false;
-          }) as WithGameState<type.State.Root, type.Move.ContributeOwnedProjects>,
+          contributeOwnedProjects: {
+            client: false,
+            move: contributeOwnedProjects,
+          },
           contributeJoinedProjects: ((G, ctx, contributions) => {
             if (!G.table.activeMoves.contributeJoinedProjects) {
               return INVALID_MOVE;
